requirements.txt
streamlit
snowflake-snowpark-python
streamlit_authenticator
streamlit-keycloak
